+++ richat/src/lib.rs
@@ -4,5 +4,6 @@ pub mod grpc;
 pub mod log;
 pub mod metrics;
 pub mod pubsub;
+pub mod richat;
 pub mod source;
 pub mod version;