+++ CHANGELOG.md
@@ -12,12 +12,22 @@ The minor version will be incremented upon a breaking change and the patch versi
 
 ### Fixes
 
+### Features
+
+### Breaking
+
+## 2025-03-20
+
+- richat-v2.4.0
+
+### Fixes
+
 - richat: do not update head on new filter with same commitment (backport of [#88](https://github.com/lamports-dev/richat/pull/88)) ([#91](https://github.com/lamports-dev/richat/pull/91))
 - richat: push messages even after SlotStatus (backport of [#89](https://github.com/lamports-dev/richat/pull/89)) ([#92](https://github.com/lamports-dev/richat/pull/92))
 
 ### Features
 
-### Breaking
+- richat: add metrics (backport of [#90](https://github.com/lamports-dev/richat/pull/90)) ([#93](https://github.com/lamports-dev/richat/pull/93))
 
 ## 2025-03-12